Split user stats and membership recalc from recalc
